@@ -22,7 +22,9 @@
     "prettier": "prettier --config .prettierrc --write \"**/*.{js,ts}\"",
     "prettier:diff": "prettier --config .prettierrc --list-different \"**/*.{js,ts}\"",
     "prettier-docs": "prettier --config .prettierrc --write \"**/*.md\"",
-    "lint": "eslint --cache \"**/*.js\" && stylelint \"**/*.css\"",
+    "lint": "yarn lint:js && yarn lint:style",
+    "lint:js": "eslint --cache \"**/*.{js,jsx,ts,tsx}\"",
+    "lint:style": "stylelint \"**/*.css\"",
     "lerna": "lerna",
     "test": "jest",
     "test:build:v2": "./admin/scripts/test-release.sh",
@@ -35,6 +37,8 @@
     "@babel/plugin-proposal-nullish-coalescing-operator": "^7.8.3",
     "@babel/plugin-proposal-optional-chaining": "^7.9.0",
     "@babel/preset-typescript": "^7.9.0",
+    "@typescript-eslint/eslint-plugin": "^3.3.0",
+    "@typescript-eslint/parser": "^3.3.0",
     "@types/express": "^4.17.2",
     "@types/fs-extra": "^8.0.1",
     "@types/inquirer": "^6.5.0",
@@ -61,14 +65,14 @@
     "concurrently": "^5.2.0",
     "enzyme": "^3.10.0",
     "enzyme-adapter-react-16": "^1.15.1",
-    "eslint": "^6.7.1",
+    "eslint": "^7.2.0",
     "eslint-config-airbnb": "^18.0.1",
-    "eslint-config-prettier": "^6.7.0",
+    "eslint-config-prettier": "^6.11.0",
     "eslint-plugin-header": "^3.0.0",
-    "eslint-plugin-import": "^2.18.2",
+    "eslint-plugin-import": "^2.21.2",
     "eslint-plugin-jsx-a11y": "^6.2.3",
-    "eslint-plugin-react": "^7.16.0",
-    "eslint-plugin-react-hooks": "^3.0.0",
+    "eslint-plugin-react": "^7.20.0",
+    "eslint-plugin-react-hooks": "^4.0.4",
     "husky": "^4.2.3",
     "jest": "^25.2.7",
     "lerna": "^3.19.0",
@@ -87,11 +91,8 @@
     "stylelint-copyright": "^2.0.0"
   },
   "lint-staged": {
-    "*.{js,jsx}": [
-      "yarn lint --fix",
-      "yarn prettier"
-    ],
-    "*.{ts,tsx}": [
+    "*.{js,jsx,ts,tsx}": [
+      "yarn eslint --fix",
       "yarn prettier"
     ],
     "*.md": [
