@@ -10,20 +10,35 @@ const WARNING = 1;
 const ERROR = 2;
 
 module.exports = {
+  root: true,
   env: {
     browser: true,
     commonjs: true,
     jest: true,
     node: true,
   },
-  parser: 'babel-eslint',
+  parser: '@typescript-eslint/parser',
   parserOptions: {
     allowImportExportEverywhere: true,
   },
   globals: {
     testStylelintRule: true,
   },
-  extends: ['airbnb', 'prettier', 'prettier/react'],
+  extends: [
+    'eslint:recommended',
+    'plugin:@typescript-eslint/eslint-recommended',
+    'plugin:@typescript-eslint/recommended',
+    'airbnb',
+    'prettier',
+    'prettier/react',
+  ],
+  settings: {
+    'import/resolver': {
+      node: {
+        extensions: ['.js', '.jsx', '.ts', '.tsx'],
+      },
+    },
+  },
   plugins: ['react-hooks', 'header'],
   rules: {
     'class-methods-use-this': OFF, // It's a way of allowing private variables.
@@ -60,6 +75,47 @@ module.exports = {
     'react/prefer-stateless-function': WARNING,
     'react/jsx-props-no-spreading': OFF,
     'react-hooks/rules-of-hooks': ERROR,
+    '@typescript-eslint/no-inferrable-types': OFF,
+    'import/first': OFF,
+    'import/order': OFF,
+    'lines-between-class-members': OFF,
+    'no-use-before-define': [
+      ERROR,
+      {functions: false, classes: false, variables: true},
+    ],
+    'no-unused-vars': OFF,
+    '@typescript-eslint/no-unused-vars': [ERROR, {argsIgnorePattern: '^_'}],
+
+    // TODO re-enable some these as errors
+    // context: https://github.com/facebook/docusaurus/pull/2949
+    '@typescript-eslint/ban-ts-comment': WARNING,
+    '@typescript-eslint/ban-types': WARNING,
+    'import/prefer-default-export': WARNING,
+    'import/no-extraneous-dependencies': WARNING,
+    'no-useless-escape': WARNING,
+    'prefer-template': WARNING,
+    'no-shadow': WARNING,
+    'no-param-reassign': WARNING,
+    'no-else-return': WARNING,
+    'no-template-curly-in-string': WARNING,
+    'array-callback-return': WARNING,
+    camelcase: WARNING,
+    'no-nested-ternary': WARNING,
+    'object-shorthand': WARNING,
+    'no-restricted-syntax': WARNING,
+    'no-unused-expressions': WARNING,
+    'consistent-return': WARNING,
+    'no-useless-return': WARNING,
+    '@typescript-eslint/no-empty-function': WARNING,
+    'global-require': WARNING,
+    'prefer-destructuring': WARNING,
+    yoda: WARNING,
+    'no-control-regex': WARNING,
+    'no-empty': WARNING,
+    'no-prototype-builtins': WARNING,
+    'no-case-declarations': WARNING,
+    'default-case': WARNING,
+    'dot-notation': WARNING,
   },
   overrides: [
     {
@@ -68,7 +124,15 @@ module.exports = {
         'packages/docusaurus-init/templates/classic/**/*.js',
       ],
       rules: {
-        'header/header': 'off',
+        'header/header': OFF,
+      },
+    },
+    {
+      files: ['*.js'],
+      rules: {
+        // Make JS code directly runnable in Node.
+        '@typescript-eslint/no-var-requires': OFF,
+        '@typescript-eslint/explicit-module-boundary-types': OFF,
       },
     },
   ],
