@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2017-present, Facebook, Inc.
+ * Copyright (c) Facebook, Inc. and its affiliates.
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
@@ -17,7 +17,7 @@ function createTestPluginContext(
   return {
     outDir: '/tmp',
     baseUrl: 'https://docusaurus.io',
-    routesPaths: routesPaths,
+    routesPaths,
     options: normalizePluginOptions(options),
   };
 }
@@ -192,7 +192,7 @@ describe('collectRedirects', () => {
                   `/def?queryString=toto`,
                 ];
               }
-              return;
+              return undefined;
             },
           },
           ['/'],
@@ -210,7 +210,7 @@ describe('collectRedirects', () => {
               if (routePath === '/') {
                 return [[`/fromPath`]] as any;
               }
-              return;
+              return undefined;
             },
           },
           ['/'],
