@@ -17,7 +17,7 @@ export default function (
   siteDir: string,
   sourceToPermalink: SourceToPermalink,
   versionedDir?: string,
-) {
+): string {
   // Determine the source dir. e.g: /website/docs, /website/versioned_docs/version-1.0.0
   let sourceDir: string | undefined;
   const thisSource = filePath;
@@ -40,7 +40,9 @@ export default function (
       if (line.trim().startsWith('```')) {
         fencedBlock = !fencedBlock;
       }
-      if (fencedBlock) return line;
+      if (fencedBlock) {
+        return line;
+      }
 
       let modifiedLine = line;
       // Replace inline-style links or reference-style links e.g:
