@@ -19,7 +19,7 @@ export function docsVersion(
   version: string | null | undefined,
   siteDir: string,
   options: PathOptions,
-) {
+): void {
   if (!version) {
     throw new Error(
       'No version tag specified!. Pass the version you wish to create as an argument. Ex: 1.0.0',
