@@ -11,7 +11,6 @@ import {MDXProvider} from '@mdx-js/react';
 import useDocusaurusContext from '@docusaurus/useDocusaurusContext';
 import renderRoutes from '@docusaurus/renderRoutes';
 import Layout from '@theme/Layout';
-import DocItem from '@theme/DocItem';
 import DocSidebar from '@theme/DocSidebar';
 import MDXComponents from '@theme/MDXComponents';
 import NotFound from '@theme/NotFound';
@@ -20,35 +19,22 @@ import {matchPath} from '@docusaurus/router';
 import styles from './styles.module.css';
 
 function DocPage(props) {
-  const {route: baseRoute, docsMetadata, location, content} = props;
+  const {route: baseRoute, docsMetadata, location} = props;
+  // case-sensitive route such as it is defined in the sidebar
+  const currentRoute =
+    baseRoute.routes.find((route) => {
+      return matchPath(location.pathname, route);
+    }) || {};
+  const {permalinkToSidebar, docsSidebars, version} = docsMetadata;
+  const sidebar = permalinkToSidebar[currentRoute.path];
   const {
-    permalinkToSidebar,
-    docsSidebars,
-    version,
-    isHomePage,
-    homePagePath,
-  } = docsMetadata;
-
-  // Get case-sensitive route such as it is defined in the sidebar.
-  const currentRoute = !isHomePage
-    ? baseRoute.routes.find((route) => {
-        return matchPath(location.pathname, route);
-      }) || {}
-    : {};
-
-  const sidebar = isHomePage
-    ? content.metadata.sidebar
-    : permalinkToSidebar[currentRoute.path];
-  const {
-    siteConfig: {themeConfig: {sidebarCollapsible = true} = {}} = {},
+    siteConfig: {themeConfig = {}} = {},
     isClient,
   } = useDocusaurusContext();
 
-  if (isHomePage) {
-    content.metadata.permalink = homePagePath;
-  }
+  const {sidebarCollapsible = true} = themeConfig;
 
-  if (!isHomePage && Object.keys(currentRoute).length === 0) {
+  if (Object.keys(currentRoute).length === 0) {
     return <NotFound {...props} />;
   }
 
@@ -59,7 +45,7 @@ function DocPage(props) {
           <div className={styles.docSidebarContainer} role="complementary">
             <DocSidebar
               docsSidebars={docsSidebars}
-              path={isHomePage ? homePagePath : currentRoute.path}
+              path={currentRoute.path}
               sidebar={sidebar}
               sidebarCollapsible={sidebarCollapsible}
             />
@@ -67,11 +53,7 @@ function DocPage(props) {
         )}
         <main className={styles.docMainContainer}>
           <MDXProvider components={MDXComponents}>
-            {isHomePage ? (
-              <DocItem content={content} />
-            ) : (
-              renderRoutes(baseRoute.routes)
-            )}
+            {renderRoutes(baseRoute.routes)}
           </MDXProvider>
         </main>
       </div>
