@@ -116,7 +116,7 @@ function NavItem({
   );
 }
 
-function MobileNavItem({items, position, className, ...props}) {
+function MobileNavItem({items, position: _position, className, ...props}) {
   // Need to destructure position from props so that it doesn't get passed on.
   const navLinkClassNames = (extraClassName, isSubList = false) =>
     clsx(
