@@ -15,7 +15,7 @@ import {matchRoutes} from 'react-router-config';
  * @param {string} pathname the route pathname, example: /docs/installation
  * @returns {Promise} Promise object represents whether pathname has been preloaded
  */
-export default function preload(routes, pathname) {
+export default function preload(routes, pathname: string) {
   const matches = matchRoutes(routes, pathname);
 
   return Promise.all(
