@@ -180,7 +180,7 @@ class CleanWebpackPlugin {
     this.removeFiles = this.removeFiles.bind(this);
   }
 
-  apply(compiler: Compiler) {
+  apply(compiler: Compiler): void {
     if (!compiler.options.output || !compiler.options.output.path) {
       // eslint-disable-next-line no-console
       console.warn(
@@ -197,7 +197,7 @@ class CleanWebpackPlugin {
      *
      * Check for hooks in-order to support old plugin system
      */
-    const hooks = compiler.hooks;
+    const {hooks} = compiler;
 
     if (this.cleanOnceBeforeBuildPatterns.length !== 0) {
       if (hooks) {
@@ -229,7 +229,7 @@ class CleanWebpackPlugin {
    *
    * Warning: It is recommended to initially clean your build directory outside of webpack to minimize unexpected behavior.
    */
-  handleInitial() {
+  handleInitial(): void {
     if (this.initialClean) {
       return;
     }
@@ -239,7 +239,7 @@ class CleanWebpackPlugin {
     this.removeFiles(this.cleanOnceBeforeBuildPatterns);
   }
 
-  handleDone(stats: Stats) {
+  handleDone(stats: Stats): void {
     /**
      * Do nothing if there is a webpack error
      */
@@ -304,7 +304,7 @@ class CleanWebpackPlugin {
     }
   }
 
-  removeFiles(patterns: string[]) {
+  removeFiles(patterns: string[]): void {
     try {
       const deleted = delSync(patterns, {
         force: this.dangerouslyAllowCleanPatternsOutsideProject,
